+++ src/components/views/search/ViewSearch.vue
@@ -8,6 +8,7 @@
             <common-input-text
                 v-model="search"
                 class="search_input"
+                height="40px"
                 placeholder="Search"
                 @update:focused="$event && ([opened = true, filtersEnabled = false])"
             >
@@ -115,7 +116,7 @@
                             <common-info-block
                                 v-for="flight in searchResults.flights.slice(0, searchLimit)"
                                 :key="flight.cid"
-                                :bottom-items="[flight.aircraft_short, flight.departure]"
+                                :bottom-items="[flight.aircraft_short?.split('/')[0], flight.departure]"
                                 is-button
                                 :top-items="[flight.callsign]"
                                 @click="['status' in flight ? mapStore.addPilotOverlay(flight.cid.toString(), true) : mapStore.addPrefileOverlay(flight.cid.toString()), opened = false]"
@@ -412,7 +413,6 @@ watch(() => mapStore.overlays.length, () => {
     @include pc {
         & &_input {
             width: 280px;
-            height: 40px;
         }
 
         &--opened .search_input {