+++ src/server/api/user/lists/[...id].ts
@@ -1,3 +1,3 @@
-import { handleListsEvent } from '~/utils/backend/lists';
+import { handleListsEvent } from '~/utils/backend/handlers/lists';
 
 export default defineEventHandler(handleListsEvent);