+++ src/store/index.ts
@@ -8,9 +8,11 @@ import type { VatsimLiveData, VatsimLiveDataShort, VatsimMandatoryData } from '~
 import { setVatsimDataStore } from '~/composables/data';
 import { useMapStore } from '~/store/map';
 import type { Coordinate } from 'ol/coordinate';
-import type { UserMapPreset, UserMapSettings } from '~/utils/backend/map-settings';
+import type { UserMapPreset, UserMapSettings } from '~/utils/backend/handlers/map-settings';
 import type { TurnsBulkReturn } from '~/server/api/data/vatsim/pilot/turns';
-import type { UserListLive, UserListLiveUser } from '~/utils/backend/lists';
+import type { UserListLive, UserListLiveUser } from '~/utils/backend/handlers/lists';
+import type { UserFilter, UserFilterPreset } from '~/utils/backend/handlers/filters';
+import type { IEngine } from 'ua-parser-js';
 
 export interface SiteConfig {
     hideSectors?: boolean;
@@ -45,6 +47,9 @@ export const useStore = defineStore('index', {
         mapSettings: {} as UserMapSettings,
         mapPresets: [] as UserMapPreset[],
         mapPresetsSaveFail: false as false | (() => Promise<any>),
+        filter: {} as UserFilter,
+        activeFilter: {} as UserFilter,
+        filterPresets: [] as UserFilterPreset[],
         config: {} as SiteConfig,
 
         showPilotStats: false,
@@ -75,8 +80,12 @@ export const useStore = defineStore('index', {
         isPC: false,
         scrollbarWidth: 0,
         device: 'desktop' as 'desktop' | 'mobile' | 'tablet',
+        engine: '' as IEngine['name'],
     }),
     getters: {
+        datalistNotSupported(): boolean {
+            return this.engine === 'Gecko';
+        },
         isTouch(): boolean {
             return this.device !== 'desktop';
         },